+++ src/procgen/PlanetGenerator.h
@@ -2,7 +2,7 @@
 
 class PlanetGenerator {
    public:
-    void generatePlanetData(std::vector<VertexAttributes> &vertexData, std::vector<uint16_t> &indices);
+    void generatePlanetData(std::vector<VertexAttributes> &vertexData, std::vector<uint32_t> &indices);
 
    private:
 };
